siemens1/src/requirements.txt
python-snap7~=1.3
setuptools==69.2.0
click==8.1.7
pydantic==2.6.4